Removing cyclical imports
Most, but not all of the cyclical imports that pylint complains about exist because
77 of the files in the repo include the phrase: import armi
So I am here trying to fix as many of the semi-cyclical imports as possible by
removing these statements.
While the code DOES run, it is probably bag practice to include the phrase almost
anywhere inside the PIP-installable ARMI package.